eval/requirements.txt
accelerate==0.23.0
albumentations==1.3.1
av==11.0.0
datasets==2.14.6
diffusers==0.25.1
docopt==0.6.2
deepspeed
easydict==1.11
einops==0.7.0
huggingface-hub==0.23.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
numpy==1.26.0
omegaconf==2.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.3
transformers
tokenizers==0.15.2
spacy
